Use the analyzer to do a rough optimization.
The idea is to find a minimal set of items according to which the
given item is "successful" in the sense of being admissible according
to the mizar analyzer.  This is a "linguistic" sense of dependence.
If the item can be verified (and not just analyzed) according to this
"analyzer-minimal" set of dependencies, then we declare that this set
is a minimal set of dependencies is "verifier" minimal.  If the
verifier dies when processing the item according to the
"analyzer-minimal" set, then we restart the process with the more
robust notation of "verifier-minimal".