RUN-IN-DIRECTORY and kin.
This still doesn't feel right to me.  The only case where we need to
set the working directory before executing a command is with the
transfer mizar tool, which (possibly) makes a directory (called
'prel') in the working directory.  Since there's apparently no option
to transfer that allows one to specify the directory in which the prel
directory, one needs to somehow have a mechanism for calling transfer
in a specific directory.  That's what exec-in-dir.sh and this new code
aims to accomplish, and it seems to work just fine.

Still, my solution feels like overkill, because all but one program
needs to actually uses the working directory.  Ideally, transfer would
accept the required command-line option; then I could just execute
programs without regard for what the working directory is, and my code
wouldn't need to be as complex as it is.  Until then, I guess I'll
live with this almost-but-not-quite-false generality.