Rewrite of ITEMIZE.
Main change: when something fails to be
accommodated/verified/exported/transferred, flag the item as a pseudo-item.

This is designed to deal primarily with the case of non-exported and
non-exportable toplevel theorems, such as those that use article-local
constants.  Consider something like:

set A = X \/ Y;
Lm: A = A;

One cannot promote the lemma to a theorem:

set A = X \/ Y;
theorem
Lm: A = A;

This is invalid because the A in the theorem is introduced via set.