Carry out our work in a fresh sandbox.
Changes the maddening interface whereby one would always
have to freshly kill the directory storing the results of itemization
before doing another round of itemization (i.e., another round of
debugging).

Use COPY-FILE from the alexandria library to move the specified
article to the sandbox.  This also deals with the problem of articles
being modified: previously, if one evaluated

  (itemize "/tmp/xboole_0.miz")

one found that the file /tmp/xboole_0.miz was modified.  This is
possibly not what was wanted; in any case, I found it annoying and
thus a problem in need of a solution.