Make a wrapper document, so that when printing we get the root of the node.
The bug here is that CXML:MAP-DOCUMENT is treating its argument as
though it were a whole document: it prints out its children.  But we
are not giving this function a document; we are giving it a child node
of the original document.  So we need to create a dummy wrapper
document, containing just the given node.