New: configuration file.
I'm so tired of pulling code down to my laptopn and then finding that
it has hard-coded paths to stuff on mizar.cs.ualberta.ca or
mws.cs.ru.nl.  The idea behind this file is that one can modify this
once and for all on a machine, never committing the changes to
it.  For this first commit, the code is set up for my laptop.