Search for just one witness, return it is found, fail otherwise.
No need to employ ALL-PATHS in ALL-PATHS-AVOID.

Try to find *one* path from SOURCE to BAD-GUY.  If there is no such
path, then all paths from SOURCE to DESTINATION avoid BAD-GUY, and we
are done.  If there is a path from SOURCE to BAD-GUY, then try to find
a path from BAD-GUY to DESTINATION.  If there is no such path, then
there is no path from SOURCE to DESTINATION that passes through
BAD-GUY, and we are done.  Otherwise, we have one path from SOURCE to
BAD-GUY and another from BAD-GUY to DESTINATION; gluing these together
gives a (counter-)witness.